Refactoring the Code
