+++ 02-DataPersistence/10-CodeAndDeployNotesAppDockerCompose/notebooks-backend/src/middlewares/validateIdHandler.js
@@ -1,0 +1,12 @@
+const mongoose = require('mongoose');
+const { AppError } = require('../appError');
+
+function validateIdHandler(req, res, next) {
+    const { id } = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return next(new AppError('Invalid notebook ID', 400));
+    }
+    next();
+}
+
+module.exports = { validateIdHandler };
