+++ 02-DataPersistence/10-CodeAndDeployNotesAppDockerCompose/notebooks-backend/src/server.js
@@ -2,7 +2,7 @@ const express = require('express');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const { notebookRouter } = require('./routes');
-const { errorHandler } = require('./errorHandler');
+const { errorHandler } = require('./middlewares/errorHandler');
 
 const app = express();
 const port = process.env.PORT || 3000;
